Re-organize data into separate modules
