+++ src/coffea/compute/data/__init__.py
@@ -1,0 +1,33 @@
+from coffea.compute.data.inputdata import InputDataGroup, InputDataset
+from coffea.compute.data.processable_data import (
+    ContextDataGroup,
+    ContextDataset,
+    ContextFile,
+    DataGroup,
+    Dataset,
+    File,
+    FileContextDataGroup,
+    StepContextDataGroup,
+    StepContextDataset,
+)
+from coffea.compute.data.rootfile import FileIterable, OpenROOTFile, ROOTFileElement
+from coffea.compute.data.step import StepElement, StepIterable
+
+__all__ = [
+    "InputDataset",
+    "InputDataGroup",
+    "ContextFile",
+    "File",
+    "ContextDataset",
+    "StepContextDataset",
+    "Dataset",
+    "ContextDataGroup",
+    "StepContextDataGroup",
+    "FileContextDataGroup",
+    "DataGroup",
+    "FileIterable",
+    "OpenROOTFile",
+    "ROOTFileElement",
+    "StepIterable",
+    "StepElement",
+]
