+++ src/coffea/compute/data/rootfile.py
@@ -1,0 +1,62 @@
+from abc import ABC, abstractmethod
+from collections.abc import Callable, Iterator
+from dataclasses import dataclass
+from typing import Generic
+
+import uproot
+from uproot import ReadOnlyDirectory
+
+from coffea.compute.context import ContextDataElement, ContextInput, Ctx_co
+from coffea.compute.data.workelement import MapData
+from coffea.compute.group import GroupedFunction
+from coffea.compute.protocol import ResultT
+
+
+@dataclass(frozen=True)
+class OpenROOTFile:
+    root_dir: ReadOnlyDirectory
+    file_path: str
+
+
+@dataclass(frozen=True)
+class ROOTFileElement:
+    """A FileElement is a lightweight representation of a file to be loaded.
+
+    It is generated by any FileIterable implementation. It is supposed to live the duration
+    of a task, but not serialized long-term. For long-term storage, use File.
+    """
+
+    path: str
+
+    def load(self) -> OpenROOTFile:
+        file = uproot.open(self.path)
+        # TODO: find a way to keep this while mocking in tests
+        # assert isinstance(file, ReadOnlyDirectory)
+        return OpenROOTFile(file, self.path)
+
+
+class FileIterable(ABC, Generic[Ctx_co]):
+    @abstractmethod
+    def iter_files(self) -> Iterator[ContextDataElement[OpenROOTFile, Ctx_co]]:
+        """Return an iterator over files in the computation."""
+        raise NotImplementedError
+
+    def map_files(self, func: Callable[[ContextInput[OpenROOTFile, Ctx_co]], ResultT]):
+        """Apply a function to each file in this data."""
+        return MapData(func=func, make_iter=self.iter_files)
+
+    def map_files_by(
+        self,
+        func: Callable[[ContextInput[OpenROOTFile, Ctx_co]], ResultT],
+        grouper: Callable[[Ctx_co], str],
+    ):
+        """Apply a function to each file in this data, grouping the output according to the key function."""
+        grouped = GroupedFunction(func=func, key_func=grouper)
+        return MapData(func=grouped, make_iter=self.iter_files)
+
+
+__all__ = [
+    "FileIterable",
+    "OpenROOTFile",
+    "ROOTFileElement",
+]
